feat(build): add --linux-categories and document the setting
The desktop entry's categories were configurable only from pyproject and
documented only in passing. Add the CLI option, matching the
action="extend" nargs="+" idiom every other list option uses, and give the
setting its own section in the publish docs with a resolution order, as
its siblings have.

Drop the `tool.flet.categories` fallback: nothing else reads it, and a
platform-neutral key would be a trap rather than a convenience, since
macOS (public.app-category.*) and Android (game/productivity) use
vocabularies in which a freedesktop value means nothing.

Also link the desktop entry's Comment and Categories keys to the
freedesktop specification, using the /desktop-entry/ path — the
/desktop-entry-spec/ form 301s to plain http.

Merge the two Linux test modules into test_build_linux.py, grouped into
classes by subject (icon staging, theme size, escaping, template
rendering), following test_project_dependencies.py and test_create.py.
