+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### New features
 
-* **App icon and desktop integration for `flet build linux`.** Linux was the one platform `flet build` shipped without an icon — `flutter_launcher_icons` has no Linux generator, so `assets/icon.png` was silently ignored and built apps ran with a generic icon. The resolved icon (`icon_linux.png`, falling back to `icon.png` or the default Flet icon) is now bundled as `data/app_icon.png` and set as the window icon on startup, which taskbars and window switchers pick up on X11 and XWayland. Because Wayland has no window-icon protocol — desktops resolve icons from an installed desktop entry matching the app id — the bundle also ships a ready-to-install freedesktop tree: `share/applications/<bundle_id>.desktop` (name from `--product`, comment from `--description`, `StartupWMClass` set) plus the icon at `share/icons/hicolor/<size>/apps/<bundle_id>.png`, and the runner now sets its program name to the bundle ID (matching upstream Flutter's runner template) so the running app maps to that entry on both Wayland and X11. See the new [App icon](https://flet.dev/docs/publish/linux#app-icon) docs ([#2269](https://github.com/flet-dev/flet/issues/2269)) by @ndonkoHenri.
+* **App icon and desktop integration for `flet build linux`.** Linux was the one platform `flet build` shipped without an icon — `flutter_launcher_icons` has no Linux generator, so `assets/icon.png` was silently ignored and built apps ran with a generic icon. The resolved icon (`icon_linux.png`, falling back to `icon.png` or the default Flet icon) is now bundled as `data/app_icon.png` and set as the window icon on startup, which taskbars and window switchers pick up on X11 and XWayland. Because Wayland has no window-icon protocol — desktops resolve icons from an installed desktop entry matching the app id — the bundle also ships a ready-to-install freedesktop tree: `share/applications/<bundle_id>.desktop` (name from `--product`, comment from `--description`, `StartupWMClass` set) plus the icon at `share/icons/hicolor/<size>/apps/<bundle_id>.png`. The entry's application categories default to `Utility` and are configurable with `--linux-categories` or `[tool.flet.linux].categories`, and the runner now sets its program name to the bundle ID (matching upstream Flutter's runner template) so the running app maps to that entry on both Wayland and X11. See the new [App icon](https://flet.dev/docs/publish/linux#app-icon) docs ([#2269](https://github.com/flet-dev/flet/issues/2269)) by @ndonkoHenri.
 
 * **macOS code signing, notarization, and Mac App Store builds in `flet build macos`.** Select a distribution lane with `--macos-distribution` (or `[tool.flet.macos.signing].distribution`): `developer-id` signs every bundled binary with your Developer ID certificate — hardened runtime, entitlements, secure timestamp — then notarizes and staples the app for direct distribution, while `app-store` produces a sandboxed app with your provisioning profile embedded, packaged into an installer-signed `.pkg` ready for App Store Connect and TestFlight. Signing identities are auto-discovered from the keychain when not explicitly configured (via CLI options, `pyproject.toml` — including per-lane `[tool.flet.macos.signing.<lane>]` subtables — or environment variables), and the whole configuration is validated before the build starts, so a typo'd identity, expired certificate, or missing store prerequisite fails in seconds instead of after the full build. See the new [Code signing](https://flet.dev/docs/publish/macos#code-signing), [Notarization](https://flet.dev/docs/publish/macos#notarization), and [Mac App Store](https://flet.dev/docs/publish/macos#mac-app-store) docs ([#2347](https://github.com/flet-dev/flet/issues/2347), [#4543](https://github.com/flet-dev/flet/issues/4543), [#6702](https://github.com/flet-dev/flet/pull/6702)) by @ndonkoHenri.
