+++ website/docs/publish/index.md
@@ -468,8 +468,8 @@ bundle_id = "com.mycompany.my_app"
 
 A short description of the application. On web builds it becomes the
 `<meta name="description">` tag and the PWA manifest's `description`; on Linux
-it becomes the `Comment` of the generated [desktop entry](linux.md#app-icon),
-shown as a tooltip in application menus. Other platforms have no equivalent
+it becomes the [`Comment`](https://specifications.freedesktop.org/desktop-entry/latest/recognized-keys.html#key-comment) of the generated
+[desktop entry](linux.md#app-icon), shown as a tooltip in application menus. Other platforms have no equivalent
 field and ignore it.
 
 #### Resolution order
@@ -492,6 +492,42 @@ flet build <target_platform> --description "Tracks your daily habits."
 ```toml
 [project]
 description = "Tracks your daily habits."
+```
+</TabItem>
+</Tabs>
+
+### Application categories
+
+:::note[Platform support]
+[Linux](linux.md) only.
+:::
+
+The [freedesktop categories](https://specifications.freedesktop.org/menu/latest/category-registry.html)
+written to the [`Categories`](https://specifications.freedesktop.org/desktop-entry/latest/recognized-keys.html#key-categories) key of the generated
+[desktop entry](linux.md#app-icon), which decide where the app is filed in
+application menus. Values must come from the registry above; unregistered ones
+are ignored by desktop environments.
+
+#### Resolution order
+
+Its value is determined in the following order of precedence:
+
+1. [`--linux-categories`](../cli/flet-build.md#--linux-categories)
+2. `[tool.flet.linux].categories`
+3. `Utility`
+
+#### Example
+
+<Tabs groupId="flet-build--pyproject-toml">
+<TabItem value="flet-build" label="flet build">
+```bash
+flet build linux --linux-categories Game Education
+```
+</TabItem>
+<TabItem value="pyproject-toml" label="pyproject.toml">
+```toml
+[tool.flet.linux]
+categories = ["Game", "Education"]
 ```
 </TabItem>
 </Tabs>
